backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "resend>=2.29.0",
]
